MDL-72991 core: Fix move_section_to is not updating the correct section
When we change the position of two sections, we just need to update the position of the affected sections, not all the sections in the course.
This will improve the performance since the system only executes the queries to affected sections.
Also, the system only clears the cache for affected sections, not all the sections in the course.