MDL-72783 usertours: Make tour name and description translatable
We can use the Moodle language format for tour's name and description now.
Also, several existing bugs were fixed in this commit when we use
the Moodle language format for step's title and content.
Now, it will display the translated text in the breadcrumb, editing page
instead of displaying the [identifier][component]