MDL-76410 general: Use mb_str_split() instead of str_split()
The use of str_split function to split a string that
needs to be split into characters (as opposed to bytes) is incorrect,
because str_split function only splits a string on a given number of bytes,
even if a character takes multiple bytes to represent.
Using the mb_str_split function is recommended, because mb_str_split
correctly splits a string accounting for multi-byte characters as well.