MDL-83005 tool_usertours: Improve the `positionBackdrop()` method
Previously, the cloned element position was relative to the parent
backdrop element. This will cause the wrong position in some cases
if the cloned element has some padding or margin. We moved the cloned
element to body, and set the position to absolute, so the cloned
element will have the same offsets as the target element